added team member data to team.json, cleaned up global styles, added logic for odd/even team component, moved all images to assets directory
