+++ src/assets/img/global/diagonal-bg.png
